test(#3515): split the harness unit tests into harness_tests.rs and name the synthetic tuple type
mod.rs was 1342 lines against the 1500-line test ceiling and round 12's
class-by-census fixes add to it, so the instrument's own unit tests move to a
sibling under graceful_shutdown_support/ (split by SUBJECT: budgets.rs pins the
deadline, mod.rs is the instrument, harness_tests.rs tests the instrument). A
child module sees its parent's private items, so the tests still reach
final_drain/collect_both_streams/CollectEnd directly.

Also fixes the clippy type_complexity error that was RED at d9cfdd98c, with a
named type alias rather than an #[allow]: the bare three-handle tuple really is
unreadable, and an allow would only hide that. Control: restoring the tuple
signature reproduces the error.
