+++ cqlite-cli/tests/graceful_shutdown_support/mod.rs
@@ -5,7 +5,8 @@
 //! module holds the INSTRUMENT — child I/O with a shared transcript, the
 //! progress-OBSERVING poll, and the bounded read-side SELECT. `budgets.rs` holds
 //! the ONE per-test deadline (round-8 descope, design.md D6a) and the unit tests
-//! that pin its invariants. The two integration tests that USE it stay in
+//! that pin its invariants; `harness_tests.rs` holds the unit tests for the
+//! instrument in THIS file. The two integration tests that USE it stay in
 //! `graceful_shutdown_tests.rs`.
 //!
 //! Included as a module (not a test target) by exactly one consumer, so the
@@ -967,376 +968,5 @@ pub fn await_write_ack(io: &ChildIo, what: &str, stage: &Stage) -> Duration {
     }
 }
 
-// ---------------------------------------------------------------------------
-// Unit coverage for the harness (the deadline's own invariants are in
-// `budgets.rs`)
-// ---------------------------------------------------------------------------
-
 #[cfg(test)]
-impl ChildIo {
-    /// A `ChildIo` with no child behind it: the returned `Sender` stands in for a
-    /// reader thread, so a unit test can make progress arrive on demand.
-    fn synthetic() -> (Self, Sender<(Stream, String)>) {
-        let (io, tx, _transcript) = Self::synthetic_with_transcript();
-        (io, tx)
-    }
-
-    /// As [`ChildIo::synthetic`], plus a handle to the shared transcript, so a
-    /// test can reproduce a reader thread that has RECORDED a line but not yet
-    /// PUBLISHED it.
-    fn synthetic_with_transcript() -> (Self, Sender<(Stream, String)>, Arc<Mutex<Vec<String>>>) {
-        let (tx, rx) = mpsc::channel();
-        let transcript: Arc<Mutex<Vec<String>>> = Arc::new(Mutex::new(Vec::new()));
-        (
-            Self {
-                rx,
-                transcript: Arc::clone(&transcript),
-            },
-            tx,
-            transcript,
-        )
-    }
-}
-
-/// OBSERVED PROGRESS MAY NOT EXTEND THE DEADLINE — the property the round-8
-/// descope exists to make true (design.md D6a).
-///
-/// Under the pre-descope design, progress on every slice reset a calibrated stall
-/// window and pushed the stage past its declared cap, which is why a declared cap
-/// was not the actual maximum. Here progress arrives on EVERY slice and the poll
-/// must still give up when the one deadline passes — and must still REPORT what it
-/// saw, because the evidence is the part worth keeping.
-///
-/// There is no timing threshold asserted: the property is that the poll
-/// TERMINATES at all under continuous progress. It is run on a worker thread with
-/// a 30s collection bound (100x the 300ms deadline under test) so a regression
-/// reports a diagnosis instead of hanging the suite.
-#[test]
-fn observed_progress_never_extends_the_deadline() {
-    let dir = tempfile::TempDir::new().expect("tempdir");
-    let data_dir = dir.path().to_path_buf();
-    let (done_tx, done_rx) = mpsc::channel();
-
-    let worker = thread::spawn(move || {
-        let (io, lines) = ChildIo::synthetic();
-        let deadline = TestDeadline::start(Duration::from_millis(300), Duration::from_millis(300));
-        let stage = deadline.stage("synthetic");
-        let outcome = poll_with_progress(&io, &data_dir, &stage, |slice, _artifacts| {
-            // Progress on every single slice.
-            let _ = lines.send((Stream::Stderr, "still working".to_string()));
-            thread::sleep(slice);
-            None::<()>
-        });
-        let report = match outcome {
-            Ok(_) => unreachable!("the step never completes"),
-            Err(fail) => fail.observed(),
-        };
-        let _ = done_tx.send(report);
-    });
-
-    let observed = done_rx.recv_timeout(Duration::from_secs(30)).expect(
-        "the progress-observing poll did not terminate within 30s despite a 300ms deadline: \
-             observed progress is extending the bound, which is exactly what the round-8 descope \
-             removed",
-    );
-    worker.join().expect("poll worker thread");
-
-    assert!(
-        observed.contains("new output line(s)"),
-        "the failure must REPORT the progress it observed: {observed}"
-    );
-    assert!(
-        observed.contains("EVIDENCE ONLY"),
-        "the failure must say the progress it observed did not extend the deadline: {observed}"
-    );
-    assert!(
-        observed.contains("ONE deadline passed"),
-        "the failure must name the one bound that ended the poll: {observed}"
-    );
-}
-
-// ---------------------------------------------------------------------------
-// EVIDENCE THAT ARRIVED BEFORE THE DEADLINE IS ACCEPTED AFTER IT LAPSES
-// (roborev job 233, finding 1 — the round-9 ruling, applied where it had only
-// been half-applied)
-// ---------------------------------------------------------------------------
-//
-// Round 9 ruled that the deadline bounds how long the test WAITS FOR EVIDENCE
-// and not whether it accepts evidence already in hand, and applied that to the
-// SUCCESS path only. The three expiry sites got it wrong in the other direction:
-// each declared a timeout without a last look at what was already queued. Under
-// contention the awaited marker, the process exit or a reader's buffer can
-// arrive well before the deadline and be consumed only after it, because this
-// thread is descheduled in between. The result was a false timeout on a working
-// product — and, worst of all, a message CONTRADICTED BY ITS OWN TRANSCRIPT.
-//
-// The three tests below each queue the evidence FIRST, then let the deadline
-// lapse, then let the harness look: each must SUCCEED.
-//
-// ON THE SLEEPS: a `sleep` can only OVERSHOOT, and overshoot makes the
-// precondition ("the deadline has already lapsed") MORE true, never less. No
-// test here asserts that anything completed FAST, so this is the opposite of the
-// #2642 wall-clock flake class.
-
-/// `wait_for`: a line that arrived before the deadline must still match.
-#[test]
-fn a_line_queued_before_the_deadline_is_matched_after_it_lapses() {
-    let (io, lines) = ChildIo::synthetic();
-    lines
-        .send((
-            Stream::Stderr,
-            format!("cqlite: {MARKER_HANDLER_ENTERED} before exit..."),
-        ))
-        .expect("queue the marker before the deadline lapses");
-
-    let deadline = TestDeadline::start(Duration::from_millis(1), Duration::from_millis(1));
-    thread::sleep(Duration::from_millis(25));
-    let stage = deadline.stage("queued-before-expiry");
-    assert!(
-        stage.remaining().is_zero(),
-        "the precondition of this test is an already-lapsed deadline"
-    );
-
-    match io.wait_for(
-        Stream::Stderr,
-        |l| l.contains(MARKER_HANDLER_ENTERED),
-        &stage,
-    ) {
-        Ok((line, _)) => assert!(line.contains(MARKER_HANDLER_ENTERED), "{line}"),
-        Err(end) => panic!(
-            "the marker had ALREADY ARRIVED when the deadline lapsed, and the wait reported \
-             {end:?} instead of matching it. That is a false timeout on a working product; on \
-             the real path it is also a self-contradicting diagnostic, because the transcript \
-             the failure prints contains the very marker the message says was never observed. \
-             (This synthetic `ChildIo` has no reader thread, so its transcript is empty by \
-             construction and is deliberately not quoted here.)"
-        ),
-    }
-}
-
-/// `poll_with_progress`: a step whose evidence landed before the deadline must
-/// still be observed, not reported as a timeout.
-#[test]
-fn a_step_completed_before_the_deadline_is_observed_after_it_lapses() {
-    let dir = tempfile::TempDir::new().expect("tempdir");
-    let data_dir = dir.path().to_path_buf();
-    // The durable artifact APPEARS before the deadline...
-    std::fs::write(data_dir.join("nb-1-big-Data.db"), b"x").expect("plant an artifact");
-
-    let (io, _lines) = ChildIo::synthetic();
-    let deadline = TestDeadline::start(Duration::from_millis(1), Duration::from_millis(1));
-    // ...and only then does the deadline lapse, before the poll ever looks.
-    thread::sleep(Duration::from_millis(25));
-    let stage = deadline.stage("queued-before-expiry");
-    assert!(
-        stage.remaining().is_zero(),
-        "the precondition of this test is an already-lapsed deadline"
-    );
-
-    let outcome = poll_with_progress(&io, &data_dir, &stage, |slice, artifacts| {
-        // The poll's OWN sample, not a scan of our own: one scan per iteration is
-        // what the documented overrun bound rests on (job 236, finding 2).
-        if artifacts >= 1 {
-            Some(())
-        } else {
-            thread::sleep(slice);
-            None
-        }
-    });
-    if let Err(fail) = outcome {
-        panic!(
-            "the artifact existed BEFORE the deadline lapsed, and the poll reported a timeout \
-             anyway — a false failure on a working product: {}",
-            fail.observed()
-        );
-    }
-}
-
-/// `collect_both_streams`: buffers delivered before the deadline must still be
-/// collected, not reported as a partial collection.
-#[test]
-fn read_side_buffers_queued_before_the_deadline_are_collected_after_it_lapses() {
-    let (tx, rx) = mpsc::channel();
-    tx.send((Stream::Stdout, b"[]".to_vec()))
-        .expect("queue stdout");
-    tx.send((Stream::Stderr, Vec::new())).expect("queue stderr");
-
-    let deadline = TestDeadline::start(Duration::from_millis(1), Duration::from_millis(1));
-    thread::sleep(Duration::from_millis(25));
-    let stage = deadline.stage("queued-before-expiry");
-    assert!(
-        stage.remaining().is_zero(),
-        "the precondition of this test is an already-lapsed deadline"
-    );
-
-    match collect_both_streams(&rx, &stage) {
-        CollectEnd::Both(out, err) => {
-            assert_eq!(
-                out, b"[]",
-                "the queued stdout buffer must be the one returned"
-            );
-            assert!(err.is_empty(), "the queued stderr buffer must be returned");
-        }
-        CollectEnd::DeadlineReached { collected } => panic!(
-            "both reader buffers were delivered BEFORE the deadline lapsed, and the collection \
-             reported a timeout with {collected}/2 collected — a false failure against a child \
-             that had already exited successfully"
-        ),
-        CollectEnd::Disconnected { collected } => panic!(
-            "unexpected disconnect with {collected}/2 collected: the senders are still alive"
-        ),
-    }
-}
-
-// ---------------------------------------------------------------------------
-// THE EXPIRY DECISION READS THE STORE THE FAILURE PRINTS FROM
-// (roborev job 236, finding 1)
-// ---------------------------------------------------------------------------
-
-/// `spawn_reader` RECORDS into the transcript before it PUBLISHES to the
-/// channel, so a reader preempted between those two operations leaves the
-/// channel behind the transcript. Job 233's expiry drain consulted only the
-/// channel, so in that window a timeout could be declared while the transcript
-/// the message prints already held the awaited marker — the self-contradicting
-/// diagnostic this change exists to prevent.
-///
-/// THE INTERLEAVING IS FORCED, NOT RACED. The plant does not sleep to arrange
-/// the ordering (which would make the precondition timing-dependent in the
-/// direction that produces flakes): the predicate itself is the synchronisation
-/// point. A decoy line — the empty line the product really does print
-/// immediately before the handler marker — is queued on the channel, and when
-/// `wait_for` receives it the predicate records BOTH lines into the transcript
-/// and returns false. So at expiry the transcript provably holds the marker and
-/// the channel provably does not, on every host and at every load.
-///
-/// The channel is left CONNECTED (this test still owns the `Sender`), so a
-/// failure here is `DeadlineReached`, not `PipesClosed`.
-#[test]
-fn a_line_recorded_but_not_yet_published_is_matched_at_expiry() {
-    let (io, lines, transcript) = ChildIo::synthetic_with_transcript();
-    let marker = format!("cqlite: {MARKER_HANDLER_ENTERED} — flushing memtable before exit...");
-
-    // The decoy the real reader publishes just before it is preempted.
-    lines
-        .send((Stream::Stderr, String::new()))
-        .expect("queue the decoy line");
-
-    let recorded = Arc::clone(&transcript);
-    let marker_for_pred = marker.clone();
-    // The reader records ONCE, as a real one does; the predicate is also applied
-    // to the transcript window at expiry, and re-recording there would say the
-    // child printed the marker twice.
-    let recorded_once = std::sync::atomic::AtomicBool::new(false);
-    let deadline = TestDeadline::start(Duration::from_millis(300), Duration::from_millis(300));
-    let stage = deadline.stage("recorded-not-published");
-
-    let matched = io.wait_for(
-        Stream::Stderr,
-        |line| {
-            if line.is_empty() && !recorded_once.swap(true, std::sync::atomic::Ordering::SeqCst) {
-                // The reader's transcript push for both lines has happened...
-                let mut t = recorded.lock().expect("transcript lock");
-                t.push("[stderr] ".to_string());
-                t.push(format!("[stderr] {marker_for_pred}"));
-                // ...and the marker's `tx.send` has NOT. The channel is now
-                // provably behind the transcript for the rest of this wait.
-            }
-            line.contains(MARKER_HANDLER_ENTERED)
-        },
-        &stage,
-    );
-
-    match matched {
-        Ok((line, _)) => assert_eq!(
-            line, marker,
-            "the matched line must be the child's own text, with the transcript's `[stderr] ` tag \
-             stripped — `await_write_ack` matches on the WHOLE trimmed line, so a tagged line \
-             would silently never match there"
-        ),
-        Err(end) => panic!(
-            "the awaited marker was IN THE TRANSCRIPT when the deadline lapsed and the wait \
-             reported {end:?} anyway. The decision read a different store from the one the \
-             failure message renders, so this failure would print the very marker it claims was \
-             never observed:\n{}\ntranscript the message would print:\n{}",
-            end.describe(),
-            io.transcript_text()
-        ),
-    }
-    drop(lines);
-}
-
-/// The window is what keeps the transcript scan from turning a CUMULATIVE store
-/// into a false pass: a line recorded BEFORE a wait began belongs to an earlier
-/// stage, which already consumed or discarded it. `await_write_ack` awaits five
-/// separate `OK`s in one test, so without the window the first would satisfy all
-/// five and a wedged session would read as green.
-#[test]
-fn a_line_recorded_before_the_wait_began_does_not_satisfy_it() {
-    let (io, lines, transcript) = ChildIo::synthetic_with_transcript();
-    transcript
-        .lock()
-        .expect("transcript lock")
-        .push("[stdout] OK".to_string());
-
-    let deadline = TestDeadline::start(Duration::from_millis(1), Duration::from_millis(1));
-    thread::sleep(Duration::from_millis(25));
-    let stage = deadline.stage("stale-transcript-line");
-    assert!(
-        stage.remaining().is_zero(),
-        "the precondition of this test is an already-lapsed deadline"
-    );
-
-    match io.wait_for(Stream::Stdout, |l| l.trim() == "OK", &stage) {
-        Ok((line, _)) => panic!(
-            "a transcript line recorded BEFORE this wait began satisfied it ({line:?}): the \
-             cumulative transcript is being read as if every line were new, so one earlier \
-             acknowledgement would satisfy every later wait for one"
-        ),
-        Err(WaitEnd::DeadlineReached { examined }) => assert_eq!(
-            examined, 0,
-            "the wait must report the size of the window it examined, and the stale line is \
-             outside it"
-        ),
-        Err(other) => panic!("expected a deadline, got {other:?}"),
-    }
-    drop(lines);
-}
-
-/// An expiry racing pipe closure must name the cause the measurement supports
-/// (roborev job 236, finding 3). Both readers have ended, so no further line can
-/// arrive: reporting `DeadlineReached` would tell the reader the pipes were
-/// "still open (more output was still possible)" about a child whose output is
-/// over — a message contradicted by the same drain that produced it.
-#[test]
-fn an_expiry_racing_pipe_closure_reports_closed_pipes() {
-    let (io, lines) = ChildIo::synthetic();
-    // A queued non-matching line, so the drain must CHECK the queue before it can
-    // observe the disconnect: `try_recv` reports `Disconnected` only once the
-    // queue is empty.
-    lines
-        .send((Stream::Stderr, "some other output".to_string()))
-        .expect("queue a non-matching line");
-    drop(lines);
-
-    let deadline = TestDeadline::start(Duration::from_millis(1), Duration::from_millis(1));
-    thread::sleep(Duration::from_millis(25));
-    let stage = deadline.stage("expiry-races-eof");
-    assert!(
-        stage.remaining().is_zero(),
-        "the precondition of this test is an already-lapsed deadline"
-    );
-
-    match io.wait_for(
-        Stream::Stderr,
-        |l| l.contains(MARKER_HANDLER_ENTERED),
-        &stage,
-    ) {
-        Err(WaitEnd::PipesClosed(_)) => {}
-        Err(WaitEnd::DeadlineReached { .. }) => panic!(
-            "both readers had ended when the deadline lapsed, and the wait reported a deadline \
-             with the pipes \"still open\" — the message names a cause its own final drain \
-             contradicts (AC2)"
-        ),
-        Ok((line, _)) => panic!("nothing matching was ever queued, yet {line:?} matched"),
-    }
-}
+mod harness_tests;
